project_01/cast-service/pyproject.toml
[tool.poetry]
name = "casts-service"
version = "0.1.0"
description = ""
authors = ["Princekrampah <jsksprince@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
databases = {extras = ["postgresql"], version = "^0.9.0"}
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
sqlalchemy = "^2.0.30"
httpx = "^0.27.0"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
